requirements.txt
viam-sdk==0.48.0

typing-extensions

pyserial
